// src/app/models/coche.models.ts
export class Coche {
  usuario?: string;
  titulo?: string;
  descripcion?: string;
  marca?:string;
  modelo?: string;
  precio?: Number;
  km?:string;
  localizacion?: string;
  combustible?: string;
  color?: string;
  puertas?: Number;
  plazas?: Number;
  cambio?: string;
  anno?: Number;
  photo?: string;
  _id?: string;
}